% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_imports.R
\docType{package}
\name{AssociationExplorer2-package}
\alias{AssociationExplorer2}
\alias{AssociationExplorer2-package}
\title{AssociationExplorer2 package}
\description{
Shiny application for exploring statistical associations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AntoineSoetewey/AssociationExplorer2}
  \item Report bugs at \url{https://github.com/AntoineSoetewey/AssociationExplorer2/issues}
}

}
\author{
\strong{Maintainer}: Antoine Soetewey \email{antoine.soetewey@uliege.be} (\href{https://orcid.org/0000-0001-8159-0804}{ORCID})

Authors:
\itemize{
  \item Cédric Heuchenne (-6196)
  \item Arnaud Claes (-1517)
  \item Antonin Descampe (-5071)
}

}
\keyword{internal}
